Several changes, restructuring how the flow works, now incorporating SQS as the component for fetch to provide events to, and database to poll from. Doing this to allow Database to process ALL of the items, instead of one/event at a time - so as to also be able to summarize updates

+++ common/dateroutines_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestCalcLongEpochFromEndDate(t *testing.T) {
 
-	expectedResult := 1609477199
+	expectedResult := int64(1609477199)
 	result := CalcLongEpochFromEndDate(2020, 12, 31)
 	assert.Equal(t, expectedResult, result)
